just did a little polishing
always add a shebang line to your executable scripts so your operating system knows how to execute your script. Then instead of typing `python3 myscript.py` you can make it executable with `chmod +x myscript.py` and execute it simply with `./myscript.py`.

In a python script you should have 2 empty lines in the top of your hierarchy inside a class or function one empty line is fine.

If you want to document your function and classes use multi-line strings in the first line after the definition and called docstrings. These are usually structured with a one line explanation of your function, followed by a longer explanation and then a formatted description of the functions parameters and return values. There are multiple standards to do that. Just choose one of them and try to stick to it. I personally like the sphinx style.

You have a lot of one line functions. I would rather write that one-liner directly instead of moving it away as a function. I'm not sure if if your also gain some performance with that but while reading the code I don't have to jump between a lot of functions to understand what it does.

Most of my edits are not my own taste but rather rules from python itself. Have a look at the PEP-8 document which describes all of that in detail. There is also a nice tool to check your code called flake8.